# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "this is an application web of interview assessment tests for interviewing out of the box .",
        "title": "test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/answers/{idTestCandidate}": {
            "post": {
                "description": "create new Answer by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Answer"
                ],
                "summary": "add new Answer",
                "parameters": [
                    {
                        "description": "Add Answer",
                        "name": "Answer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Answer"
                        }
                    },
                    {
                        "type": "string",
                        "description": "idTestCandidate",
                        "name": "idTestCandidate",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Answer"
                        }
                    }
                }
            }
        },
...
        "/questions": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "find a question by type or difficulty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "find a question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question search by type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "question search by difficulty",
                        "name": "difficulty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Question"
                            }
                        }
                    }
                }
            }
        },
        "/questions/edit": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "create new question by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "add new  question",
                "parameters": [
                    {
                        "description": "Add question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateQuestionInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                }
            }
        },
        "/questions/edit/{id}": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "find a question to edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "find a question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "update a question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "question"
                ],
                "summary": "update a question",
                "parameters": [
                    {
                        "description": "Update question",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateQuestionInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Question"
                        }
                    }
                }
            }
        },
...
        "/skill": {
            "post": {
                "description": "create new skill by json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skill"
                ],
                "summary": "add new  skill",
                "parameters": [
                    {
                        "description": "Add Skill",
                        "name": "skill",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSkillInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Skill"
                        }
                    }
                }
            }
        },
        "/skills": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "get all skills",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "skill"
                ],
                "summary": "get skills",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SkillsResponse"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Answer": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "answer_choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnswerChoices"
                    }
                },
                "answer_file": {
                    "type": "string"
                },
                "answer_text": {
                    "type": "string"
                },
                "candidate_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "point": {
                    "type": "number"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.AnswerChoices": {
            "type": "object",
            "required": [
                "choices_id"
            ],
            "properties": {
                "answer_id": {
                    "type": "integer"
                },
                "choices_id": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Candidate": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                }
            }
        },
...
        "models.Choices": {
            "type": "object",
            "required": [
                "choice_text",
                "is_answer"
            ],
            "properties": {
                "choice_text": {
                    "type": "string"
                },
                "is_answer": {
                    "type": "boolean"
                }
            }
        },
...
        "models.CreateQuestionInput": {
            "type": "object",
            "required": [
                "difficulty",
                "expected_time",
                "max_points",
                "name",
                "question_text"
            ],
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choices"
                    }
                },
                "difficulty": {
                    "type": "string"
                },
                "expected_time": {
                    "type": "integer"
                },
                "file_read_me": {
                    "type": "string"
                },
                "max_points": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "skill_id": {
                    "type": "integer"
                },
                "skill_name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.CreateSkillInput": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Question": {
            "type": "object",
            "required": [
                "name",
                "question_text"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Answer"
                    }
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Choices"
                    }
                },
                "difficulty": {
                    "type": "string"
                },
                "expected_time": {
                    "type": "integer"
                },
                "file_read_me": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "max_points": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "skill": {
                    "$ref": "#/definitions/models.Skill"
                },
                "skill_id": {
                    "type": "integer"
                },
                "test": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Test"
                    }
                },
                "test_questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TestQuestion"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Skill": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SkillsResponse": {
            "type": "object",
            "required": [
                "ID",
                "name"
            ],
            "properties": {
                "ID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Test": {
            "type": "object",
            "properties": {
                "archived": {
                    "type": "boolean"
                },
                "candidate": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Candidate"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "notify_emails": {
                    "type": "string"
                },
                "passing_score": {
                    "type": "integer"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Question"
                    }
                },
                "show_score": {
                    "type": "boolean"
                },
                "time_limit": {
                    "type": "integer"
                },
                "timing_policy": {
                    "type": "string"
                }
            }
        },
...
        "models.TestQuestion": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "test_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}